+++ inc/libft/ft_btree.h
@@ -6,12 +6,13 @@
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/11 15:30:29 by asoursou          #+#    #+#             */
-/*   Updated: 2020/06/02 19:13:17 by asoursou         ###   ########.fr       */
+/*   Updated: 2020/06/21 02:06:06 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_BTREE_H
 # define FT_BTREE_H
+# define FT_REQUIRE_TYPE_BOOL
 # define FT_REQUIRE_TYPE_GENERIC
 # define FT_REQUIRE_TYPE_SIZE_T
 # include "private/ft_include.h"
@@ -49,7 +50,20 @@ void	ft_btree_infix(t_btree *root, t_gfunction function);
 t_btree	*ft_btree_insert(t_btree **root, t_btree *element, t_gcompare cmp);
 
 /*
+** Returns true if the given node is a leaf.
+*/
+t_bool	ft_btree_is_leaf(t_btree *node);
+
+/*
+** Merge root with its left and right child.
+** The left and right pointer of the root are overwritten by the new values.
+** Returns root.
+*/
+t_btree	*ft_btree_merge(t_btree *root, t_btree *left, t_btree *right);
+
+/*
 ** Creates a new element with malloc.
+** Returns NULL on error.
 */
 t_btree	*ft_btree_new(void *content);