+++ inc/libft/ft_memory.h
@@ -24,6 +24,7 @@ void	ft_bzero(void *s, size_t n);
 ** Allocates enough space for count objects that are size bytes of memory each
 ** and returns a pointer to the allocated memory.
 ** The allocated memory is filled with bytes of value zero.
+** Returns NULL on error.
 */
 void	*ft_calloc(size_t count, size_t size);