+++ src/btree/ft_btree_is_leaf.c
@@ -1,23 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_btree_new.c                                     :+:      :+:    :+:   */
+/*   ft_btree_is_leaf.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/11/11 15:42:11 by asoursou          #+#    #+#             */
-/*   Updated: 2020/06/21 02:10:45 by asoursou         ###   ########.fr       */
+/*   Created: 2020/06/21 02:07:02 by asoursou          #+#    #+#             */
+/*   Updated: 2020/06/23 17:34:40 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "ft_btree.h"
 
-t_btree	*ft_btree_new(void *content)
+t_bool	ft_btree_is_leaf(t_btree *node)
 {
-	t_btree *b;
-
-	if ((b = (t_btree*)malloc(sizeof(t_btree))))
-		ft_btree_merge(b, NULL, NULL)->content = content;
-	return (b);
+	return (!(node->left || node->right));
 }