+++ src/btree/ft_btree_merge.c
@@ -1,23 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_btree_new.c                                     :+:      :+:    :+:   */
+/*   ft_btree_merge.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: asoursou <asoursou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2019/11/11 15:42:11 by asoursou          #+#    #+#             */
-/*   Updated: 2020/06/21 02:10:45 by asoursou         ###   ########.fr       */
+/*   Created: 2020/06/21 02:00:34 by asoursou          #+#    #+#             */
+/*   Updated: 2020/06/21 02:01:31 by asoursou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <stdlib.h>
 #include "ft_btree.h"
 
-t_btree	*ft_btree_new(void *content)
+t_btree	*ft_btree_merge(t_btree *root, t_btree *left, t_btree *right)
 {
-	t_btree *b;
-
-	if ((b = (t_btree*)malloc(sizeof(t_btree))))
-		ft_btree_merge(b, NULL, NULL)->content = content;
-	return (b);
+	root->left = left;
+	root->right = right;
+	return (root);
 }